+++ application/config/routes.php
@@ -49,7 +49,12 @@ defined('BASEPATH') OR exit('No direct script access allowed');
 | Examples:	my-controller/index	-> my_controller/index
 |		my-controller/my-method	-> my_controller/my_method
 */
-$route['default_controller'] = 'welcome';
+$route['default_controller'] = 'hogwarts';
 $route['404_override'] = '';
 $route['translate_uri_dashes'] = FALSE;
-$route['lock/em/up'] = 'welcome/shucks';
+$route['lock/em/up'] = 'hogwarts/shucks';
+
+$route['sleep'] = 'First/zzz';
+
+/*How to Fix It	Remapped using wildcard (show / digit)*/
+$route['shw(:num)'] = 'First/gimme/$1';